relocate submission unit tests
